7709 - Migrate Shared.Data.Documents.Abstractions into the Infrastructure/Abstractions/ and remove all references to that project
