+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.API.sln
@@ -57,8 +57,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers.csproj", "{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj", "{7B827585-4F7A-473B-B994-681BE96B0C04}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Core\xxENSONOxx.xxSTACKSxx.Shared.Core.csproj", "{3F133394-CF28-4110-BA3D-EBF404208A81}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core.Tests", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Core.Tests\xxENSONOxx.xxSTACKSxx.Shared.Core.Tests.csproj", "{06DE9A76-034B-4975-8365-A7F88EF7954E}"
@@ -345,18 +343,6 @@ Global
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x64.Build.0 = Release|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x86.ActiveCfg = Release|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|x86.Build.0 = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|x64.Build.0 = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|x86.Build.0 = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|x64.ActiveCfg = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|x64.Build.0 = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|x86.ActiveCfg = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|x86.Build.0 = Release|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|x64.ActiveCfg = Debug|Any CPU
