+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Application.QueryHandlers/SearchMenuQueryHandler.cs
@@ -3,13 +3,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using Amazon.DynamoDBv2.DocumentModel;
 using xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Queries;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions;
 #if (DynamoDb)
 using Amazon.DynamoDBv2.DocumentModel;
 using xxENSONOxx.xxSTACKSxx.Infrastructure.Abstractions;
 #endif
 using xxENSONOxx.xxSTACKSxx.CQRS.Queries.SearchMenu;
 using xxENSONOxx.xxSTACKSxx.Domain;
+using xxENSONOxx.xxSTACKSxx.Infrastructure.Abstractions;
 
 namespace xxENSONOxx.xxSTACKSxx.Application.QueryHandlers;
