+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure.UnitTests/DynamoDbMenuRepositoryTests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents;
 using FluentAssertions;
 using NSubstitute;
 using Xunit;
