+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/Abstractions/CosmosOperationResult.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents;
 
 namespace xxENSONOxx.xxSTACKSxx.Infrastructure.Abstractions
 {
