+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/Abstractions/IDynamoDbObjectStorage.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents;
 
 namespace xxENSONOxx.xxSTACKSxx.Infrastructure.Abstractions;
