+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/HealthChecks/CustomHealthCheck.cs
@@ -1,6 +1,5 @@
 using System.Threading;
 using System.Threading.Tasks;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using xxENSONOxx.xxSTACKSxx.Domain;
 //using xxENSONOxx.xxSTACKSxx.Domain.Entities;
@@ -13,10 +12,10 @@ namespace xxENSONOxx.xxSTACKSxx.Infrastructure.HealthChecks;
 /// Resources implementing IHealthCheck should be added directly to the pipeline:
 /// i.e: services.AddHealthChecks().AddCheck<CosmosDbDocumentStorage<Menu>>("CosmosDB");
 /// </summary>
-public class CustomHealthCheck(IDocumentStorage<Menu> documentStorage) : IHealthCheck
+public class CustomHealthCheck(Abstractions.IDocumentStorage<Menu> documentStorage) : IHealthCheck
 {
     //Example: do not add this example, CosmosDBStorage already implements IHealthCheck
-    readonly IDocumentStorage<Menu> documentStorage = documentStorage;
+    readonly Abstractions.IDocumentStorage<Menu> documentStorage = documentStorage;
 
     public async Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
     {
