+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/Repositories/CosmosDbMenuRepository.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Threading.Tasks;
-using xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions;
 using xxENSONOxx.xxSTACKSxx.Application.Integration;
 using xxENSONOxx.xxSTACKSxx.Domain;
 
 namespace xxENSONOxx.xxSTACKSxx.Infrastructure.Repositories;
 
-public class CosmosDbMenuRepository(IDocumentStorage<Menu> documentStorage) : IMenuRepository
+public class CosmosDbMenuRepository(Abstractions.IDocumentStorage<Menu> documentStorage) : IMenuRepository
 {
     public async Task<Menu> GetByIdAsync(Guid id)
     {
