+++ src/func-aeh-listener/src/functions/xxENSONOxx.xxSTACKSxx.Listener.sln
@@ -15,8 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests.csproj", "{6D545555-2426-4478-82A9-D756A00F1E39}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj", "{7B827585-4F7A-473B-B994-681BE96B0C04}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Core\xxENSONOxx.xxSTACKSxx.Shared.Core.csproj", "{3F133394-CF28-4110-BA3D-EBF404208A81}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj", "{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}"
@@ -53,10 +51,6 @@ Global
 		{6D545555-2426-4478-82A9-D756A00F1E39}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D545555-2426-4478-82A9-D756A00F1E39}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D545555-2426-4478-82A9-D756A00F1E39}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -84,7 +78,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
 		{6D545555-2426-4478-82A9-D756A00F1E39} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
-		{7B827585-4F7A-473B-B994-681BE96B0C04} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
 		{3F133394-CF28-4110-BA3D-EBF404208A81} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
 		{B26B7E36-8E4F-45EE-BEBA-02262F9DCDC6} = {EFD81450-4A57-43C5-BBFF-76DAF9EABB70}
