+++ src/func-cosmosdb-worker/.template.config/template.json
@@ -28,7 +28,6 @@
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Configuration.Tests/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Core/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Core.Tests/**",
-				"shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.DependencyInjection/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus.Tests/**",
