+++ src/func-cosmosdb-worker/src/functions/xxENSONOxx.xxSTACKSxx.Worker.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers\xxENSONOxx.xxSTACKSxx.Shared.Messaging.Handlers.csproj", "{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj", "{7B827585-4F7A-473B-B994-681BE96B0C04}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Core\xxENSONOxx.xxSTACKSxx.Shared.Core.csproj", "{3F133394-CF28-4110-BA3D-EBF404208A81}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Core.Tests", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Core.Tests\xxENSONOxx.xxSTACKSxx.Shared.Core.Tests.csproj", "{06DE9A76-034B-4975-8365-A7F88EF7954E}"
@@ -75,10 +73,6 @@ Global
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B827585-4F7A-473B-B994-681BE96B0C04}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3F133394-CF28-4110-BA3D-EBF404208A81}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -117,7 +111,6 @@ Global
 		{7290B15D-B0DB-48FA-9A2A-1E8E15447CD2} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{51D0C76D-9F7F-4213-9BC1-0D7DE060E4AF} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{6C460E51-98B2-4DD7-B4FC-C25C6C3F739E} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
-		{7B827585-4F7A-473B-B994-681BE96B0C04} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{3F133394-CF28-4110-BA3D-EBF404208A81} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{06DE9A76-034B-4975-8365-A7F88EF7954E} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
