--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions/IDocumentStorage.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-
-namespace xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions
-{
-    public interface IDocumentStorage<TEntity> where TEntity : class
-    {
-        Task<OperationResult<TEntity>> SaveAsync(string identifier, string partitionKey, TEntity document, string eTag);
-        Task<OperationResult<TEntity>> GetByIdAsync(string identifier, string partitionKey);
-        Task<OperationResult> DeleteAsync(string identifier, string partitionKey);
-    }
-}
