--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <Description>Data Abstraction package for common features used throughout the projects.</Description>
-    <Company>Ensono</Company>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/ensono/stacks-dotnet-packages-data-abstractions</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/ensono/stacks-dotnet-packages-data-abstractions</RepositoryUrl>
-    <RepositoryType>GIT</RepositoryType>
-    <LangVersion>12.0</LangVersion>
-  </PropertyGroup>
-
-</Project>
